backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn",
    "supabase",
    "python-dotenv",
    "google-generativeai>=0.8.5",
    "pydantic-ai>=1.27.0",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.20",
    "psycopg2-binary>=2.9.11",
    "logfire>=4.16.0",
    "nest-asyncio>=1.6.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
